comment delete, update finished. users auth done.

+++ app/Http/Controllers/ArticleController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -34,7 +36,10 @@ class ArticleController extends Controller
 
     public function add()
     {
-        return view("articles.add");
+        $categories = Category::all();
+        return view("articles.add", [
+            "categories" => $categories
+        ]);
     }
 
     public function create()
@@ -53,6 +58,7 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
         $article->save();
 
         return redirect("/articles");
@@ -60,8 +66,11 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
+        if(Gate::denies('post-delete', $article)) {
+            return back()->with('article-delete-error', "Article delete failed");
+        }
 
-        return redirect("/")->with("info", "An article deleted");
+        $article->delete();
+        return redirect("/")->with("article-delete-success", "An article deleted");
     }
 }
