+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -17,6 +18,7 @@ class CommentController extends Controller
         $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -25,8 +27,33 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back();
+        if (Gate::denies('comment-permission', $comment)) {
+            return back()->with('comment-delete-error', 'Unauthorize');
+        }
+
+        $comment->delete();
+        return back()->with('comment-delete-success', 'comment deleted');
+    }
+
+
+    public function edit ($id)
+    {
+        $comment = Comment::find($id);
+        return view('articles/comment-edit', [
+            "comment" => $comment
+        ]);
+    }
+    public function update($id)
+    {
+        $comment = Comment::find($id);
+
+        if (Gate::denies('comment-permission', $comment)) {
+            return back()->with('comment-update-error', 'Unauthorize');
+        }
+
+        $comment->content = request()->content;
+        $comment->save();
+        return redirect("/articles/detail/$comment->article_id")->with('comment-update-success', 'comment-updated');
     }
 }
